Refactor App.tsx into AppContent component; enhance alert and file data synchronization hooks
- Decomposed App.tsx into AppContent for improved modularity and maintainability.
- Introduced useAlertsFlow and useFileDataSync hooks for better alert management and file data handling.
- Updated tests to reflect changes in navigation lock and mock hooks.
- Enhanced performance metrics documentation with recent profiling data.
- Updated code review documentation to reflect architectural changes and performance opportunities.
