+++ docs/development/performance/2025-10-07-baseline.md
@@ -49,7 +49,7 @@ console.table(tracker.getRenderProfiles());
 
 ## Immediate observations
 
-- Main bundle remains ~491 kB raw / 133 kB gzip. Splitting orchestration out of `App.tsx` into feature-specific entry points should be the first optimization story.
+- Main bundle remains ~491 kB raw / 133 kB gzip. Splitting orchestration out of `AppContent` into feature-specific entry points should be the first optimization story.
 - `utils-*` is the largest non-core chunk; start by auditing Autosave/File storage exports for code that can move behind dynamic imports.
 - Navigation timings are already captured—monitor future runs for spikes when filesystem prompts appear or large case lists load.
